feat: add WorkflowUserSettings component for managing user settings in workflows
