+++ web/components/interface/workflow-details-modal.tsx
@@ -4,10 +4,7 @@ import Icon from "@/components/misc/icon";
 import { useWorkflowRuns } from "@/lib/hooks/use-workflow-runs";
 import { fetchAPI } from "@/lib/pulse-editor-website/backend";
 import { Workflow, WorkflowRun } from "@/lib/types";
-import {
-  deleteWorkflowSetting,
-  setWorkflowSetting,
-} from "@/lib/workflow-settings";
+import WorkflowUserSettings from "@/components/interface/workflow-user-settings";
 import {
   addToast,
   Button,
@@ -27,7 +24,6 @@ import {
   Tooltip,
 } from "@heroui/react";
 import { useState } from "react";
-import useSWR from "swr";
 
 function CopyRow({ label, value }: { label: string; value: string }) {
   return (
@@ -519,108 +515,3 @@ function WorkflowRunHistory({ workflowId }: { workflowId: string }) {
   );
 }
 
-function WorkflowUserSettings({ workflowId }: { workflowId: string }) {
-  const { data: settings, mutate } = useSWR<Record<string, string>>(
-    `/api/workflow/user-settings/get?workflowId=${encodeURIComponent(workflowId)}`,
-  );
-
-  const [newKey, setNewKey] = useState("");
-  const [newValue, setNewValue] = useState("");
-  const [newIsSecret, setNewIsSecret] = useState(false);
-
-  return (
-    <div className="flex flex-col gap-y-2">
-      <p className="text-sm font-semibold">User Settings</p>
-      <p className="text-default-500 text-xs">
-        These settings apply to all apps in this workflow. Per-app settings take
-        priority over workflow-level settings.
-      </p>
-      {settings && Object.keys(settings).length > 0 && (
-        <div className="space-y-1">
-          {Object.entries(settings).map(([key, value]) => (
-            <WorkflowSettingInput
-              key={key}
-              workflowId={workflowId}
-              setting={{ key, value }}
-              onUpdated={() => mutate()}
-            />
-          ))}
-        </div>
-      )}
-      <div className="flex items-center gap-x-1">
-        <Input label="Key" size="sm" value={newKey} onValueChange={setNewKey} />
-        <Input
-          label="Value"
-          size="sm"
-          value={newValue}
-          onValueChange={setNewValue}
-          type={newIsSecret ? "password" : "text"}
-        />
-        <Button
-          isIconOnly
-          variant="light"
-          onPress={() => setNewIsSecret((prev) => !prev)}
-        >
-          {newIsSecret ? <Icon name="lock" /> : <Icon name="lock_open" />}
-        </Button>
-        <Button
-          isIconOnly
-          variant="light"
-          isDisabled={newKey.trim() === ""}
-          onPress={async () => {
-            await setWorkflowSetting(
-              workflowId,
-              newKey.trim(),
-              newValue,
-              newIsSecret,
-            );
-            mutate();
-            setNewKey("");
-            setNewValue("");
-            setNewIsSecret(false);
-          }}
-        >
-          <Icon name="add" />
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function WorkflowSettingInput({
-  workflowId,
-  setting,
-  onUpdated,
-}: {
-  workflowId: string;
-  setting: { key: string; value: string };
-  onUpdated: () => void;
-}) {
-  const isRedacted = setting.value === "redacted";
-
-  return (
-    <div className="flex items-center gap-x-1">
-      <Input
-        label={setting.key}
-        size="sm"
-        value={setting.value}
-        isReadOnly
-        type={isRedacted ? "password" : "text"}
-      />
-      <Tooltip content="Delete setting">
-        <Button
-          isIconOnly
-          variant="light"
-          color="danger"
-          size="sm"
-          onPress={async () => {
-            await deleteWorkflowSetting(workflowId, setting.key);
-            onUpdated();
-          }}
-        >
-          <Icon name="delete" />
-        </Button>
-      </Tooltip>
-    </div>
-  );
-}
